requirements.txt
Flask==2.0.2
PyJWT==2.3.0
Werkzeug==2.0.2
Flask-CORS==3.0.10
